Refactor aiRouter in ai.ts to enhance chat functionality
- Introduced comprehensive toolset for managing Gmail interactions, including listing threads, marking threads as read/unread, and creating labels.
- Implemented robust request validation using Zod for improved error handling.
- Streamlined connection management and integrated database queries for user authorization.
- Removed redundant code from ZeroAgent class in chat.ts, consolidating chat message handling within aiRouter.
